Looks good to me!

+++ app/controllers/pets_controller.rb
@@ -16,15 +16,24 @@ class PetsController < ApplicationController
 
   def update
     pet = Pet.find(params[:pet_id])
-    pet.update(pet_params)
-    pet.save
-    redirect_to "/pets/#{pet.id}"
+    if pet.update(pet_params)
+      redirect_to "/pets/#{pet.id}"
+    else 
+      flash[:notice] = pet.errors.full_messages.to_sentence
+      redirect_to "/pets/#{pet.id}/edit"
+    end
   end
 
   def destroy
     pet = Pet.find(params[:pet_id])
-    pet.destroy
-    redirect_to "/pets"
+    if pet.approved_application == nil
+      session[:favorites].delete(params[:pet_id]) if session[:favorites] != nil
+      pet.destroy
+      redirect_to "/pets"
+    else 
+      flash[:notice] = "Cannot delete a pet with an active adoption application!"
+      redirect_to "/pets"
+    end
   end
 
   private
